Add footer actions to PageDialog and make proposal save button sticky
